# dsp-integration/docs/dsp-integration/api/openapi.yaml
openapi: 3.1.0
info:
  title: Personalisation Hub — Display Types & DSP Integration API
  version: 1.0.0-draft
  description: |
    The agreed contract for every API added by the Display Types & DSP
    Integration build. Implement exactly these endpoints, fields and error
    codes; do not add endpoints, fields or behaviour that are not here.
    Changes to this file are made in the same commit as the code change.

    Two surfaces:
      * Partner API (/v1)       — DSPs and direct advertisers (tier 2).
      * Admin API (/admin/v1)   — behind HQ Admin screens.
    Plus the public sellers.json. The OpenRTB bid request/response contract
    is documented in API.md (it is sent by us, not served).

    Out of scope (existing Personalisation Hub functionality, no endpoints
    here): playback, playlists' playback behaviour, targeting evaluation,
    distribution to players, playback logging, campaign playback analytics.

    Jobs with no endpoint: the SSP auction runs as a scheduled job
    (CLI for demos: npm run auction:run); billing line items are stored
    only (CLI to print them: npm run billing:print). Neither has a UI,
    report or API.
servers:
  - url: https://{instance}/api
    variables:
      instance:
        default: hub.example-retailer.com
        description: The retailer's own Personalisation Hub instance (in its VPC).
...
security:
  - partnerToken: []
paths:
...
  /v1/reservations:
    post:
      tags: [Partner — Reservations]
      operationId: createReservation
      summary: Reserve (named advertiser) or bid (CPM) for a play window
      description: Approved and activated campaigns only, while the window's auction is open (from auctionOpensHours before the auction cutoff until the cutoff; Advertiser settings → Auction schedule). Pre-auction checks apply (floor, lists, categories, approval, activation).
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReservationCreate'}
      responses:
        '201': {description: Accepted, content: {application/json: {schema: {$ref: '#/components/schemas/Reservation'}}}}
        '400': {$ref: '#/components/responses/ValidationFailed'}
        '401': {$ref: '#/components/responses/Unauthorised'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/NotEligible'}
  /v1/reservations/{reservationId}:
    get:
      tags: [Partner — Reservations]
      operationId: getReservation
      summary: One reservation or bid, with its outcome
      parameters: [{name: reservationId, in: path, required: true, schema: {type: string}}]
      responses:
        '200': {description: Reservation, content: {application/json: {schema: {$ref: '#/components/schemas/Reservation'}}}}
        '401': {$ref: '#/components/responses/Unauthorised'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '404': {$ref: '#/components/responses/NotFound'}

  # ------------------------------------------------------------------ Admin
...
components:
  securitySchemes:
    partnerToken:
      type: http
      scheme: bearer
      description: Token issued to a connected DSP or tier-2 partner. Identifies the partner; advertiser is passed per request.
    adminSession:
      type: apiKey
      in: cookie
      name: ph_session
      description: |
        Existing HQ Admin session. Who sees what (spec, Who sees each section):
        admin users get everything, including DSP Integration and editing
        advertiser settings; marketing users get Display Types, Playlist
        Management, Advertisers / Inventory (read-only on the per-advertiser
        settings) and Campaign Status; help desk users get none of it.
        Scopes: admin (DSP Integration, saving advertiser settings), approver
        (approve/reject; HQ Admin role by default, open question 39),
        sections (the four shared sections). In the POC there is no cookie —
        the session is a stand-in whose role comes from the POC_ROLE env var
        (hq_admin | hq_marketing | hq_helpdesk).
...
  responses:
    Unauthorised: {description: Missing or invalid credentials, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    NotFound: {description: 'Not found, or not visible to this caller', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    ValidationFailed: {description: 'error.code = validation_failed; details lists each field', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    NotEligible: {description: 'error.code = not_approved | below_floor | advertiser_blocked | category_blocked | not_on_whitelist | not_invited | targeting_not_supported', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Conflict: {description: 'error.code = conflict (e.g. version changed, already submitted, window sold, bidding not open or closed, live not allowed)', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    RateLimited: {description: 'error.code = rate_limited. Per partner token (default 50 requests/s, bursts of 100), or too many asset uploads in flight at once. Retry-After gives the seconds to wait.', headers: {Retry-After: {schema: {type: integer}}}, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [validation_failed, variable_not_permitted, checks_failed, not_approved, below_floor, advertiser_blocked, category_blocked, not_on_whitelist, not_invited, targeting_not_supported, conflict, has_dependents, unauthorised, forbidden, not_found, rate_limited, internal_error]
            message: {type: string}
            details:
              type: array
              items:
                type: object
                properties:
                  field: {type: string}
                  variable: {type: string}
                  reason: {type: string}
...
    ReservationCreate:
      type: object
      required: [positionId, windowStart, campaignId, advertiserId, type, bidCpm]
      properties:
        positionId: {type: string}
        windowStart: {type: string, format: date-time}
        campaignId: {type: string}
        advertiserId: {type: string}
        type: {type: string, enum: [reserve, bid]}
        bidCpm: {type: number, description: 'The CPM, in the company currency: the bid (type bid), or the reservation price agreed through the DSP (type reserve). Must clear the effective floor. A reservation is booked at this price.'}
    Reservation:
      type: object
      required: [reservationId, status]
      properties:
        reservationId: {type: string}
        status: {type: string, enum: [pending, won, lost, reserved, rejected]}
        clearingCpm: {type: [number, 'null']}
        currency: {type: string}
        reason: {type: [string, 'null']}
...